+++ src/main/java/com/misterd/agritechtwo/gui/custom/RaisedBedBlockScreen.java
@@ -37,7 +37,7 @@ public class RaisedBedBlockScreen extends AbstractContainerScreen<RaisedBedBlock
                 this.leftPos, this.topPos, 0.0F, 0.0F,
                 this.imageWidth, this.imageHeight, 256, 256);
 
-        float growthProgress = this.menu.blockEntity.getGrowthProgress();
+        float growthProgress = this.menu.getGrowthProgress() / 100.0F;
         if (growthProgress > 0.0F) {
             int barHeight = (int) (18 * growthProgress);
             int barY = this.topPos + 18 + 18 - barHeight;
@@ -54,7 +54,7 @@ public class RaisedBedBlockScreen extends AbstractContainerScreen<RaisedBedBlock
     protected void extractTooltip(GuiGraphicsExtractor graphics, int mouseX, int mouseY) {
         if (mouseX >= this.leftPos + 122 && mouseX <= this.leftPos + 128
                 && mouseY >= this.topPos + 18 && mouseY <= this.topPos + 37) {
-            float progress = this.menu.blockEntity.getGrowthProgress();
+            float progress = this.menu.getGrowthProgress() / 100.0F;
             graphics.setComponentTooltipForNextFrame(this.font, List.of(
                     Component.translatable("tooltip.agritechtwo.growth_progress"),
                     Component.literal(String.format("%.1f%%", progress * 100.0F))